{#2,#3,#4,#6} refacoring and documentation
