+++ app/src/main/java/de/teamawesome/awesomeplayer/fragments/listFragments/AlbumsListFragment.java
@@ -1,0 +1,24 @@
+package de.teamawesome.awesomeplayer.listFragments;
+
+import static de.teamawesome.awesomeplayer.listFragments.ListUtils.*;
+
+import android.os.Bundle;
+import android.view.View;
+import android.widget.ListView;
+
+/**
+ * A {@link CursorListFragment} used to display albums.
+ */
+
+public class AlbumsListFragment extends de.teamawesome.awesomeplayer.listFragments.CursorListFragment {
+    /**
+     * Used to trigger the Frgment transition to a {@link MediaListFragment} displaying all songs
+     * from the clicked album.
+     */
+    @Override
+    public void onListItemClick(ListView l, View v, int position, long id) {
+        Bundle arguments = de.teamawesome.awesomeplayer.listFragments.ListBundles.MEDIA_FROM_ALBUM_BUNDLE.get();
+        arguments.putStringArray(SELECTION_ARGS, new String[]{"" + id});
+        fragmentListener.onFragmentInteraction(arguments, this);
+    }
+}
