+++ app/src/main/java/de/teamawesome/awesomeplayer/fragments/listFragments/ListBundles.java
@@ -8,10 +8,14 @@ import android.os.Bundle;
 import java.io.Serializable;
 
 /**
- * Created by sven on 6/5/17.
+ * Specifies a bunch of Bundles which can be accessed via the .get() method.
+ * These Bundles can be used to quickly set up the different List views.
  */
 
 public enum ListBundles {
+    /**
+     * used to display all albums in a {@link AlbumsListFragment}
+     */
     ALBUM_BUNDLE ( new Object[][]{
         {URI, ALBUMS_URI},
         {FROM, ALBUMS_FROM},
@@ -20,6 +24,9 @@ public enum ListBundles {
         {SELECTION_ARGS, null}
     }),
 
+    /**
+     * used to display all songs in a {@link MediaListFragment}
+     */
     MEDIA_BUNDLE ( new Object[][]{
         {URI, MEDIA_URI},
         {FROM, MEDIA_FROM},
@@ -28,14 +35,23 @@ public enum ListBundles {
         {SELECTION_ARGS, null}
     }),
 
+    /**
+     * Used to display all songs from a specific album(s) in a {@link MediaListFragment}.
+     * It is required to set the {@link ListUtils}.SELECTION_ARGS key's value to an
+     * array holding the album's id!
+     */
     MEDIA_FROM_ALBUM_BUNDLE ( new Object[][]{
             {URI, MEDIA_URI},
             {FROM, MEDIA_FROM},
             {PROJECTION, null},
             {SELECTION_STRING, MEDIA_FROM_ALBUM_SELECTION_STRING},
-            {SELECTION_ARGS, null} //Should be set to an array containing the clicked album's is
+            {SELECTION_ARGS, null} //Should be set to an array containing the clicked album's id
     }),
 
+    /**
+     * Used to display all songs from a specific playlist in a {@link MediaListFragment}.
+     * It is required to set the {@link ListUtils}.URI key's value to an string like MediaStore.Audio.Playlists.Members.getContentUri("external", #playlistid#).toString()!
+     */
     MEDIA_FROM_PLAYLIST_BUNDLE ( new Object[][]{
             {URI, null}, //Should be set to the correct uri from the clicked playlist's id
             {FROM, MEDIA_FROM},
@@ -44,7 +60,9 @@ public enum ListBundles {
             {SELECTION_ARGS, null}
     }),
 
-
+    /**
+     * used to display all playlists in a {@link PlaylistsListFragment}
+     */
     PLAYLIST_BUNDLE ( new Object[][]{
         {URI, PLAYLISTS_URI},
         {FROM, PLAYLISTS_FROM},
@@ -68,6 +86,9 @@ public enum ListBundles {
         }
     }
 
+    /**
+     * @return Returns the {@link Bundle} associated with the enum key.
+     */
     public Bundle get(){
         return new Bundle(associatedBundle);
     };
