+++ app/src/main/java/de/teamawesome/awesomeplayer/fragments/listFragments/MediaListFragment.java
@@ -1,0 +1,8 @@
+package de.teamawesome.awesomeplayer.listFragments;
+
+/**
+ * A {@link CursorListFragment} used to display songs.
+ */
+
+public class MediaListFragment extends CursorListFragment {
+}
