+++ app/src/main/java/de/teamawesome/awesomeplayer/fragments/listFragments/PlaylistsListFragment.java
@@ -8,15 +8,19 @@ import android.widget.ListView;
 import static de.teamawesome.awesomeplayer.listFragments.ListUtils.URI;
 
 /**
- * Created by sven on 6/5/17.
+ * A {@link CursorListFragment} used to display playlists.
  */
 
-public class PlaylistsListFragment extends MediaStoreListFragment {
+public class PlaylistsListFragment extends CursorListFragment {
+    /**
+     * Used to trigger the Fragment transition to a {@link MediaListFragment} displaying all songs
+     * from the clicked playlist.
+     */
     @Override
     public void onListItemClick(ListView l, View v, int position, long id) {
         Bundle arguments = ListBundles.MEDIA_FROM_PLAYLIST_BUNDLE.get();
         arguments.putString(URI, MediaStore.Audio.Playlists.Members.getContentUri("external", id).toString());
-        fListener.onFragmentInteraction(arguments, this);
+        fragmentListener.onFragmentInteraction(arguments, this);
     }
 
 }
