--- app/src/main/java/de/teamawesome/awesomeplayer/listFragments/MediaListFragment.java
@@ -1,8 +1,0 @@
-package de.teamawesome.awesomeplayer.listFragments;
-
-/**
- * Created by sven on 6/5/17.
- */
-
-public class MediaListFragment extends MediaStoreListFragment {
-}
